==> notebooks/0 INDEX.ipynb <==
# Index


1. Before we start... [Setup necessary python and libs](1 Setup.ipynb)

2. Jupyter notebooks  
  2.1. [A Short intro](2.1 Short Intro to jupyter notebooks.ipynb)  
  2.2. Notes about performance [Random Numbers](2.2 Playing with Random - Performance.ipynb)  

3. Digging into data  
  3.1. Data setup [Generating Sample Data](3.1 Generating Data.ipynb)  
  3.2. Log extraction  
    * Challenge 1 [Extract From Log File](3.2.A Challenge 1 - Extract From Log File.ipynb)
    * [Solution](3.2.B Challenge 1 - Solution.ipynb)  

  -- break --  

  3.3. Grouping Data  
    * Challenge 2 [Group Information](3.3.A Challenge 2 - Top 10 channels with most activity.ipynb)  
    * [Solution](3.3.B Challenge 2 - Solution.ipynb)  

  3.4. Joining Data  
    * Challenge 3 [Join DataFrames](3.4.A Challenge 3 - Joining - Names of Top 10 channels.ipynb)  
    * [Solution](3.4.B Challenge 3 - Solution.ipynb)  

  3.5. Recommendations  
    * Challenge 4 [Who Watched X also watched Y](3.5.A Challenge 4 - Recommendations.ipynb)  
    * [Solution](3.5.B Challenge 4 - Solution.ipynb)
