==> notebooks/1 Setup.ipynb <==
# Before we start...

Download this notebook from https://github.com/ajscarvalho/pandas_workshop

it will be needed later to create a random log file so we can begin playing with data...

# Setup (debian)

# install latest python (3.6), pip3 and virtalenv

```
sudo apt-get install python3 python3.6 python3-pip
sudo pip3 install --upgrade pip
sudo pip3 install virtualenv
```



# Create a project and an environment 

###### (paths are up to you)

```
cd <project_dir>
mkdir <env>
virtualenv -p /usr/bin/python3.6 <env>
```

e.g.
```
    project_dir = 'pandas_workshop'
    env = 'pandas_workshop_env'
```



# Activate the environment

```
source <env>/bin/activate
```

###### You just entered in your environment

Now packages you install will be just for this environment.
Also pip and python will refer to the version you declared with virtualenv.

## check it!

pip -V



# install and run python notebooks
```
pip install jupyter
mkdir notebooks
jupyter notebook --port 9999 --notebook-dir=notebooks [--no-browser]
```

You should now have an interactive online coding session, essential for experimentation.



# Install libs for this workshop
```
pip install numpy pandas
```
<font color=red>install dask?</font>
