==> notebooks/2 Short Intro to jupyter notebooks.ipynb <==
# Get familiar with python notebooks

- create a new python 3 notebook
- experiment a bit running some python commands
- give a name to the notebook
- save it
- create cells with different variables
- cells can be run in any order, but usually executed in sequence


<br><br><br><br>
## A Short intro into magic jupyter commands

%magic -brief


%timeit?

import time
time.sleep(0.5)

%timeit time.sleep(0.1)

%%timeit
time.sleep(0.5)

<br><br><br><br>
## Let's start by playing with Random Number Generation

import random

def get_random_numbers(minimum, maximum, amount):
    numbers = []
    for i in range(amount):
        numbers.append(random.randint(minimum, maximum))
        
    return numbers


print('Test:')
print ("Random Int[0-10]:", random.randint(0, 10))
print ("5 Random Int[0-10]:", get_random_numbers(0, 10, 5))

%timeit random.randint(0, 10)


%timeit get_random_numbers(0, 10, 100)


<br>
<br>
<br>
<br>
# 100 Times Slower as expected. But...
<br>
<br>
<br>
<br>

## Enters numpy, the lib for number operations in python
<br>


import numpy as np

%timeit np.random.randint(0, 10)

### 1 single run - not much of a performance improvement... but let's get 100 random integers

print(np.random.randint(0, 10, 100))

%timeit np.random.randint(0, 10, 100)

### Now that's what I call an improvement! 100 random numbers take barelly more than 2x as long
<br>
<br>
#### Also there's a lot more random number generator functions that may be useful 

integersAccordingToStatisticalDistribution = [int(np.floor(x)) for x in np.random.exponential(100, 100)]
print(integersAccordingToStatisticalDistribution)
print('Mean is', np.mean(integersAccordingToStatisticalDistribution))
