==> notebooks/3.1 Generating Data.ipynb <==
# Generating Data

Let's get going...

We start with no data, but let's generate some fake data...

As I'm currently working at NOS, let's generate some random EPG Data, and then a fake Activity Log

with your active environment, run the following scripts in bin/

```   
python bin/channel_generator.py -s channelCount   # generates channels                     [default     100   records]
python bin/customer_base_generator.py -s users    # generates user base                    [default   10K     records]
python bin/epg_program_generator.py -s programs   # generates program metadata             [default   10K     records]
python bin/epg_event_generator.py -s days         # generates epg grid (for each Channel)  [default      10      days]
python bin/log_generator.py -r rubbishPct -s size # generates an activity log with noise   [default 3M        records] 
                                                                                           [default      50 % rubbish]
```

This will create files on the dwh directory inside the project folder.

The last script takes about 5 minutes to run and generates a ~600MB File, so reduce the size, if you wish, depending on your machine
... and don't forget to erase the files
