==> notebooks/3.2.A Challenge 1 - Extract From Log File.ipynb <==
<br><br>
# Challenge 1 - Extract

Extract Meaningful Data from the Log, preferably in the most efficient way 

Take every ChangeChannel Action and Write it to a CSV file


### Cheat Sheet

File Location: will depend on current location


#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')
activityFilename = os.path.join(dwhdir, 'activity.log')

print(activityFilename)


#### read file line by line:


import io

with io.open(filename, "r", encoding='utf-8') as file:
    for line in file:
        do_something(line)

#### Lines we're looking for: ChangeChannel

**2017-01-02 00:41:28**:carlina.langton.depletion User@**9d587747555e4d19b4915b11c18c276a**: UserAction=**ChangeChannel**(**C-0008**, **E/00904859**, **C/57752**) chickening!always#pestilence


(1) Timestamp <br>
(2) UserId <br>
(3) Action <br>
(4) Channel <br>
(5) Current Event Id <br>
(6) Some Other Id <br>

#### a regexp to catch the content of the line

import re

line = "2017-01-02 00:41:28:carlina.langton.depletion User@9d587747555e4d19b4915b11c18c276a: UserAction=ChangeChannel(C-0008, E/00904859, C/57752) chickening!always#pestilence"

timestampRE = '^.([\d\s:-]+)'
userIdRE = 'User@([\w]+)'
userActionRE = 'UserAction=([\w]+)'
changeChannelRE = 'UserAction=ChangeChannel\(([\w-]+), ([\w/]+),'

print (re.search(timestampRE, line).group(1))
print (re.search(userIdRE, line).group(1))
print (re.search(userActionRE, line).group(1))
print (re.search(changeChannelRE, line).group(1))
print (re.search(changeChannelRE, line).group(2))

### Saving in CSV?

Using python pandas (at last)


#### Create a DataFrame from a list

```
import pandas as pd

df = pd.DataFrame([[1,2], [3,4]], columns=['A', 'B'])
```

#### Write dataframe to CSV

```
df.to_csv('filename', sep=',', encoding='utf-8', index=False)
```

#### Reading Dataframe from CSV

```
df.read_csv('filename', sep=',', encoding='utf-8', header=False, dtypes=[str, str, int, ...])
```
