==> notebooks/3.2.B Challenge 1 - Solution.ipynb <==
#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')

activityFilename  = os.path.join(dwhdir, 'activity.log')
dataframeFilename = os.path.join(dwhdir, 'channel_changes.csv')

print('In :', activityFilename)
print('Out:', dataframeFilename)

#other imports
import io
import re
import pandas as pd





def get_good_lines(filename, lineFilterStr):
    data = []
    with io.open(filename, "r", encoding='utf-8') as file:
        for line in file:
            isMatch = filter_line(line, lineFilterStr)
            if isMatch:
                data.append(line)
    
    return data







def filter_line(line, filterStr):
    return filterStr in line




from datetime import datetime # another way to measure

startTS = datetime.now()
changeChannelLines = get_good_lines(activityFilename, 'Action=ChangeChannel')

print('Process took',  (datetime.now() - startTS).total_seconds(), 'seconds')

print(len(changeChannelLines), 'lines') # about 1M lines extracted from a 184MB file in 1.5s

#now we can apply regexps - compile them beforehand
userRE = re.compile('User@([\w]+)')
changeChannelRE = re.compile('UserAction=ChangeChannel\(([\w-]+), ([\w/]+),')

def extract_features(line):
    ts = line[0:19]
    user = userRE.search(line).group(1)
    changeChannel = changeChannelRE.search(line)
    channel = changeChannel.group(1)
    event = changeChannel.group(2)
    return (ts, user, channel, event)
    
#Test it
print (extract_features(changeChannelLines[0]))


#extract everything to a table (of sorts)
startTS = datetime.now()

datatable = [extract_features(l) for l in changeChannelLines]

print('Extracting features took',  (datetime.now() - startTS).total_seconds(), 'seconds')

#create a dataframe

startTS = datetime.now()

df = pd.DataFrame(datatable, columns=['Timestamp', 'User', 'Channel', 'Event'])

print('Dataframe Creation took',  (datetime.now() - startTS).total_seconds(), 'seconds')

print(df.head(5))


startTS = datetime.now()

df.to_csv(dataframeFilename, sep=',', encoding='utf-8', index=False)

print('CSV Writting took', (datetime.now() - startTS).total_seconds(), 'seconds')
