==> notebooks/3.3.A Challenge 2 - Top 10 channels with most activity.ipynb <==
<br><br>
# Challenge 2 - Group

analyse data extracted from the log to determine the channel with the most activity

Write conclusions to a CSV file


### Cheat Sheet

File Location: will depend on current location


#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')
channelChangesFile = os.path.join(dwhdir, 'channel_changes.csv')

print(channelChangesFile)


### Create/ Read / Write Dataframe (recap)

#### Create a DataFrame from a list

```
import pandas as pd

df = pd.DataFrame([[1,2], [3,4]], columns=['A', 'B'])
```

#### Write dataframe to CSV

```
df.to_csv('filename', sep=',', encoding='utf-8', index=False)
```

#### Reading Dataframe from CSV

```
df.read_csv('filename', sep=',', encoding='utf-8', header=False, dtypes=[str, str, int, ...])
```


### Quickly Inspecting Dataframe
`df.head(numberOfLines)`

`df.tail(numberOfLines)`


### Grouping data in dataframe

import pandas
pandas.DataFrame.group*?

pandas.DataFrame.groupby?

#this is example code, not meant to be run,...

df.groupby(by='colA')
df.groupby(by=['colA', 'colB'], sort=False)

# and aggregate
df.groupby(by=['colA', 'colB'], sort=False).count()

df.groupby(by=['colA', 'colB'], sort=False).agg(count, sum, mean)

#print the result, it may not be what you actually want...




## Channel Changes Dataframe sample
```
I            Timestamp                              User Channel       Event  
0  2017-01-02 00:00:00  7f98c243497046a086b09a84d68aa0f1  C-0016  E/00904866
```

### Column Selection

`df = df[['Channel', 'User']]`

#### or if you prefer:

```
cols = ['Channel', 'User']
df = df[cols]
```

### Filtering:

```
filter = (df.Channel == 'C-0016')
c0016_activity = df[filter]

# in pandas we use the bitwise operator to compose expressions {&, |, ~}
filter = (df.Channel == 'C-0016') & (User == 'me') 

ChannelWhitelist = ['C-0001', 'C-0002', 'C-0003']
filter = df.Channel.isin(ChannelWhitelist) # any listlike thing
```

## Let's Try it out

#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')

channelChangesFilename = os.path.join(dwhdir, 'channel_changes.csv')

print('In :', channelChangesFilename)

import pandas as pd

df = pd.read_csv(channelChangesFilename)
print (df.head())

print (df.shape)

f1 = df[ df.Channel == 'C-0016' ]

print (f1.shape)
print (f1.head())


ChannelWhitelist = ['C-0016', 'C-0031', 'C-0026']
filterChannel = df.Channel.isin(ChannelWhitelist)
c1_to_3 = df[filterChannel]

print (c1_to_3.shape)
print (c1_to_3.head())
