==> notebooks/3.3.B Challenge 2 - Solution.ipynb <==
#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')

channelChangesFilename = os.path.join(dwhdir, 'channel_changes.csv')
channelChangesCountFilename = os.path.join(dwhdir, 'channel_changes_count.csv')

print('In :', channelChangesFilename)
print('Out:', channelChangesCountFilename)

#other imports
import pandas as pd





def read_channel_changes_csv(filename):
    df = pd.read_csv(filename)
    return df
#    print(df.head())
    



channelChanges = read_channel_changes_csv(channelChangesFilename)



def group_then_count(df, col, sort=True):
    dfGroup = df[[col]].copy()
    dfGroup['Count'] = 0
    dfCount = dfGroup.groupby(by=col, sort=sort).count()
    return dfCount


print ( group_then_count(channelChanges, 'Channel'       ).head(10)     )
print ( group_then_count(channelChanges, 'Channel', False).head(10)     )
print ( group_then_count(channelChanges, 'Channel', False).nlargest(10, 'Count') )

%timeit group_then_count(channelChanges, 'Channel').head(10)
%timeit group_then_count(channelChanges, 'Channel', False).head(10)
%timeit group_then_count(channelChanges, 'Channel', False).nlargest(10, 'Count')


# Save dataframe
df = group_then_count(channelChanges, 'Channel')
df.reset_index(inplace=True) # Channel index is reinserted as a column and index is recreated from 0
df.to_csv(channelChangesCountFilename, sep=',', encoding='utf-8', index=False)
