==> notebooks/3.4.B Challenge 3 - Solution.ipynb <==
#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')

channelChangesCountFilename = os.path.join(dwhdir, 'channel_changes_count.csv')
channelsRefDataFilename     = os.path.join(dwhdir, 'channels.csv')

topChannelsFilename         = os.path.join(dwhdir, 'top_channels.csv')

print('In :', channelChangesCountFilename)
print('In :', channelsRefDataFilename)

print('Out:', topChannelsFilename)

#other imports
import pandas as pd
pd.options.display.max_rows = 10
pd.options.display.max_colwidth = 180
pd.options.display.width = 180



def read_channel_changes_count_csv(filename):
    df = pd.read_csv(filename)
    return df


def read_channels_ref_data_csv(filename):
    df = pd.read_csv(filename, names=['Channel', 'Position', 'Definition', 'Abbrev', 'Name'], dtype={'Position': int} )
    return df




topChannelChanges = read_channel_changes_count_csv(channelChangesCountFilename)
channelsRefData   = read_channels_ref_data_csv    (channelsRefDataFilename)

print (topChannelChanges.shape[0], "\n", topChannelChanges.head(1))
print (channelsRefData.shape[0],   "\n", channelsRefData.head(1))


def join_on_col(dfA, dfB, col):
    dfC = dfA.merge(dfB, on=col, how='inner')
    return dfC



topChannels = join_on_col(topChannelChanges, channelsRefData, 'Channel')
print (topChannels.head(5))

print()

topChannels = join_on_col(channelsRefData, topChannelChanges, 'Channel')
print (topChannels.head(5))

# Save dataframe
topChannels.to_csv(topChannelsFilename, sep=',', encoding='utf-8', index=False)
