==> notebooks/3.5.B Challenge 4 - Solution.ipynb <==
#relative paths
import os, sys, inspect

currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parentdir  = os.path.dirname(currentdir)

rootdir = parentdir # is it?
dwhdir = os.path.join(rootdir, 'dwh')

channelChangesFilename      = os.path.join(dwhdir, 'channel_changes.csv')
eventsRefDataFilename       = os.path.join(dwhdir, 'events.csv')
programsRefDataFilename     = os.path.join(dwhdir, 'programs.csv')


print('In :', channelChangesFilename)
print('In :', eventsRefDataFilename)
print('In :', programsRefDataFilename)


#other imports
from datetime import datetime
import pandas as pd

pd.options.display.max_rows = 10
pd.options.display.max_colwidth = 180
pd.options.display.width = 180

#useful functions
class Timer(object):

    def __init__(self):
        self.lastTime = None

    def time(self, msg="Time passed", restart=False):
        if restart: self.lastTime = None

        now = datetime.now()
        if self.lastTime:
            print ("{:40s}: {:7.3f}s".format(msg, (now - self.lastTime).total_seconds()))
        
        self.lastTime = now

perfTimer = Timer()




def read_channel_changes_csv(filename):
    df = pd.read_csv(filename,
        names=['Timestamp', 'UserId', 'Channel', 'EventId'],
        header=0 # when passing names, it's assumed the start of the file as the start of data... 
                 # file already has headers in row 0 
    )
    #cut information that we will not need
    df = df[['UserId', 'EventId']].copy()
    df.drop_duplicates(inplace=True)
    return df


def read_events_ref_data_csv(filename):
    df = pd.read_csv(filename,
        names=['EventId', 'ChannelId', 'ProgramId', 'Duration', 'StartDate', 'EndDate'], 
        dtype={'Position': int} 
    )
    return df


def read_programs_ref_data_csv(filename):
    df = pd.read_csv(filename, dtype={'SeriesId': str, 'SeriesSeason': str, 'SeriesEpisode': str})
    # after looking and checking there are NaN's:
    df = df.fillna('')
    return df



perfTimer.time(restart=True)

channelChanges  = read_channel_changes_csv  ( channelChangesFilename)

perfTimer.time("Reading Channel Changes")

eventsRefData   = read_events_ref_data_csv  (  eventsRefDataFilename)

perfTimer.time("Reading Events Ref Data")

programsRefData = read_programs_ref_data_csv(programsRefDataFilename)

perfTimer.time("Reading Programs Ref Data")

print ('\n\nCHANNEL CHANGES:',  channelChanges.shape[0], "\n",  channelChanges.head(1))
print ('\n\nEVENTS         :',   eventsRefData.shape[0], "\n",   eventsRefData.head(1))
print ('\n\nPROGRAMS       :', programsRefData.shape[0], "\n", programsRefData.head(1))


def join_on_col(dfA, dfB, col):
    dfC = dfA.merge(dfB, on=col, how='inner')
    return dfC




def get_recommendations(userId, programId):

    eventsForProgram         = get_events_for_program(programId)
    usersForProgram          = get_users_for_events(eventsForProgram)
    eventIdsSeenByUsers      = get_event_ids_for_users(usersForProgram)
    otherEventIdsSeenByUsers = remove_event_ids_of_program(eventIdsSeenByUsers, eventsForProgram)
    otherEventsSeenByUsers   = join_with_events(otherEventIdsSeenByUsers)
    programsSeenByUsers      = get_programs_for_users(otherEventsSeenByUsers, limit=40)
    
    return programsSeenByUsers


#get a randomUser()

def get_random_user():
    return channelChanges.UserId.sample().iloc[0]

def get_random_program(): # to guarantee that the event actually exists
    change      = channelChanges.sample().iloc[0].to_dict()
    eventId     = change.get('EventId')
    programId   = eventsRefData[eventsRefData.EventId == eventId].iloc[0].ProgramId
    program     = programsRefData[programsRefData.ProgramId == programId] 
    return program.iloc[0].to_dict()



userId = get_random_user()
program = get_random_program()

print("User", userId)
print("Program", program)

programId = program.get('ProgramId')


def get_events_for_program(programId):
    events = eventsRefData[eventsRefData.ProgramId == programId]
    return list(events.to_dict()['EventId'].values())

print ("Program", programId, type(programId))

perfTimer.time(restart=True)
eventsForProgram = get_events_for_program(programId)
perfTimer.time("TIMING: events for program")

print ("Found {} events for that program {}".format(len(eventsForProgram), programId))



def get_users_for_events(eventList):
    print('eventList', type(eventList), eventList)
    changes = channelChanges[channelChanges.EventId.isin(eventList)].copy()
    print ('get_users_for_events:', changes.shape[0], 'changes')
    users = changes.UserId.unique()
    return users

eventsForProgram = get_events_for_program(programId)

perfTimer.time(restart=True)
usersForProgram = get_users_for_events(eventsForProgram)
perfTimer.time("TIMING: get users for program")

print(len(usersForProgram), 'users found: e.g. ', usersForProgram[0:5])

def get_event_ids_for_users(userList):
    
    userActivity = channelChanges[channelChanges.UserId.isin(userList)].copy()
    print("Total Activity of Users {:7d}".format(userActivity.shape[0]) )
    #unqUserActivity = unqUserActivity.drop_duplicates(['UserId', 'EventId']) # already did it at the source
    activityByEvent = userActivity.groupby(by='EventId', sort=False).count()
    activityByEvent.columns = ['Count'] # rename remaining column (UserId) that will be used to store count to Count
    print('Total Distinct Events   {:7d}'.format(activityByEvent.shape[0]), '\n', activityByEvent.tail(1))
    return activityByEvent

programEvents     = get_events_for_program(programId)
usersForProgram   = get_users_for_events(programEvents)

perfTimer.time(restart=True)
eventsSeenByUsers = get_event_ids_for_users(usersForProgram)
perfTimer.time("TIMING: get events for users")

print ("Completelly unnecessary: Top Events: ")
eventsSortedByTopAudience = eventsSeenByUsers.sort_values(by='Count', ascending=False)
print(eventsSortedByTopAudience.head(2))


def remove_event_ids_of_program(eventDf, eventsToRemove):
    print (eventDf.loc['E/00905067'])
    
    # try to filter on index with not in - time it
    #events = eventDf[ ~eventDf.EventId.isin(eventsToRemove) ] # must filter on index

    #alternative - use column instead of index
    eventsNoIndex = eventDf.reset_index()
    events = eventsNoIndex[ ~eventsNoIndex.EventId.isin(eventsToRemove) ] # must filter on index
    
    return events

eventsForProgram  = get_events_for_program(programId)
usersForProgram   = get_users_for_events(eventsForProgram)
eventsSeenByUsers = get_event_ids_for_users(usersForProgram)

perfTimer.time(restart=True)
otherEvents       = remove_event_ids_of_program(eventsSeenByUsers, eventsForProgram)
perfTimer.time("TIMING: get events for users")

print ("From {} events, we're removing {} eventIds and stayed with {} events".format(
    eventsSeenByUsers.shape[0], len(eventsForProgram), otherEvents.shape[0])
)


def join_with_events(eventIdAudience):
    return join_on_col(eventIdAudience, eventsRefData, 'EventId')

eventsForProgram         = get_events_for_program(programId)
usersForProgram          = get_users_for_events(eventsForProgram)
eventIdsSeenByUsers      = get_event_ids_for_users(usersForProgram)
otherEventIdsSeenByUsers = remove_event_ids_of_program(eventIdsSeenByUsers, eventsForProgram)

perfTimer.time(restart=True)
otherEventsSeenByUsers   = join_with_events(otherEventIdsSeenByUsers)
perfTimer.time("TIMING: get events for users")

print("X-Check", otherEventIdsSeenByUsers.shape[0], 'to', otherEventsSeenByUsers.shape[0])
print("Dataframe Check\n", otherEventsSeenByUsers.head(4))


def get_programs_for_users(events, limit=None):
    #join with events to get programId
    summary = events[['ProgramId', 'Count']]
    programs = summary.groupby('ProgramId', sort=False).sum()
    if limit:
        programs.sort_values(by="Count", ascending=False, inplace=True)
        programs = programs.head(limit)
    
    programs = pd.merge(
        left=programs,         left_index=True,
        right=programsRefData, right_on='ProgramId'
    )
    
    if not limit:
        programs.sort_values(by="Count", ascending=False, inplace=True)
        
    return programs

eventsForProgram         = get_events_for_program(programId)
usersForProgram          = get_users_for_events(eventsForProgram)
eventIdsSeenByUsers      = get_event_ids_for_users(usersForProgram)
otherEventIdsSeenByUsers = remove_event_ids_of_program(eventIdsSeenByUsers, eventsForProgram)
otherEventsSeenByUsers   = join_with_events(otherEventIdsSeenByUsers)

perfTimer.time(restart=True)
programsSeenByUsers      = get_programs_for_users(otherEventsSeenByUsers)
perfTimer.time("TIMING: get all programs for users")

print(programsSeenByUsers.shape[0], 'are all programs')

perfTimer.time(restart=True)
programsSeenByUsers      = get_programs_for_users(otherEventsSeenByUsers, limit=40)
perfTimer.time("TIMING: get some programs for users")

print(programsSeenByUsers.shape[0], 'are some programs')



%timeit get_programs_for_users(otherEventsSeenByUsers)



%timeit get_programs_for_users(otherEventsSeenByUsers, limit=40)


programs = get_programs_for_users(otherEventsSeenByUsers)
print(programs[['ProgramId','Count']].head(40)) #oops - needs sort

programs = get_programs_for_users(otherEventsSeenByUsers, limit=40)
print(programs[['ProgramId','Count']].head(40))


perfTimer.time(restart=True)

eventsForProgram         = get_events_for_program(programId)
usersForProgram          = get_users_for_events(eventsForProgram)
eventIdsSeenByUsers      = get_event_ids_for_users(usersForProgram)
otherEventIdsSeenByUsers = remove_event_ids_of_program(eventIdsSeenByUsers, eventsForProgram)
otherEventsSeenByUsers   = join_with_events(otherEventIdsSeenByUsers)
programsSeenByUsers      = get_programs_for_users(otherEventsSeenByUsers, limit=40)

perfTimer.time("TIMING: recommendation")
